Restructure project data
The code is no longer unorganized, but structured into subfolders that allow installation via setuptools

Resolves #3
