+++ paw/tests/test_files/test_pw_all_std.txt
